chore(endereco): mover tipo Endereco para o redux slice e stringify-address para lib, atualizar imports
