+++ frontend/src/pages/enderecos-cliente.tsx
@@ -3,7 +3,7 @@ import { PageTitle } from "@/components/page-title";
 import { Card } from "@/components/ui/card"; 
 import { Table } from "@/components/ui/table";
 import { ArrowLeftIcon } from "@/components/icons/arrow-left-icon";
-import { stringifyAddress, type Endereco } from "@/consts/enderecos";
+import { stringifyAddress } from "@/lib/stringify-address";
 import { Separator } from "@/components/ui/separator";
 import { Button } from "@/components/ui/button";
 import { ArrowRightIcon } from "@/components/icons/arrow-right-icon";
@@ -14,7 +14,7 @@ import { useEffect } from "react";
 import { useSelector } from "react-redux";
 import type { RootState } from "@/redux/root-reducer";
 import { selectClienteById } from "@/redux/clientes/slice";
-import { selectAllEnderecos } from "@/redux/endereco/slice";
+import { selectAllEnderecos, type Endereco } from "@/redux/endereco/slice";
 import { useDispatch } from "react-redux";
 import type { AppDispatch } from "@/redux/store";
 import { fetchEnderecos } from "@/redux/endereco/fetch";
