+++ frontend/src/pages/reservas-cliente.tsx
@@ -15,7 +15,7 @@ import { useNavigate, useParams, useLocation } from "react-router"
 import { useDispatch, useSelector } from "react-redux"
 import type { RootState } from "@/redux/root-reducer"
 import { Separator } from "@/components/ui/separator"
-import { stringifyAddress } from "@/consts/enderecos"
+import { stringifyAddress } from "@/lib/stringify-address"
 import type { AppDispatch } from "@/redux/store"
 import { fetchReservas } from "@/redux/reservas/fetch"
 import { selectClienteById } from "@/redux/clientes/slice"
